refactor: share SCCM duplicate-key JSON preflight
